Moved Facia to use the DogpileAction.
While in there also shifted some of the boilerplate into its own file to get it out of the way of the working code.
